Massive code overhaul.
In an attempt to speed up performance (as well as optimize code),
I have switched over to a lookup table for stuff in the Skopik class.

It gets generated once just like CharUtils, and should hopefully
increase performance when used correctly.

Of course, this never works out in my favor, and performance gradually
gets worse with each iteration. I'll figure it out eventually...

Other changes include (but are not limited to):

- Added support for binary numbers
- Added support for unsigned-long numbers
- Added support for hexadecimal explicit array indices
- Added support for negative hexadecimal/binary numbers
- Added some more helper methods in the Skopik class (see above notes)
- Reworked the data types enum to allow for more detailed information