Overall improvement of both code quality and performance.
I finally solved the performance problem...it wasn't much to do with
my tokenizer code or any of the more "obvious" problems...oh no...

It was the test file being loaded in for parsing...EVERY ITERATION.

It's an absolute miracle my parser "worked" in the first place!

There were a lot of changes in this commit, so I can't remember
every little change/fix/improvement I made. Mostly it was me focusing
on the tokenizer code, thinking the problem was there.

Word of advice:
USE THE VISUAL STUDIO PROFILER! Create a detailed report!
I spent _hours_ trying to find the bottleneck in my code, but 5 minutes
in the profiler and I found it straight away.